backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "openai>=2.7.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "roboflow>=1.2.11",
    "uvicorn>=0.38.0",
]
